refactor: Split isAuthed check that was common to the resolvers into separate function; Move JWT functions out from services for semantic and import cycle reasons
