+++ api/app/services/signin_service.go
@@ -2,6 +2,7 @@ package services
 
 import (
 	"app"
+	"app/auth"
 	"app/ent"
 	"app/ent/user"
 	"context"
@@ -24,7 +25,7 @@ func (s *signinService) Signin(ctx context.Context, email string, password strin
 		return nil, fmt.Errorf("password incorrect: %w", err)
 	}
 
-	token, err := New(s.client).GenerateToken(user.ScreenName)
+	token, err := auth.GenerateToken(user.ScreenName)
 	if err != nil {
 		return nil, err
 	}
