+++ api/app/services/signin_service_test.go
@@ -1,6 +1,7 @@
 package services
 
 import (
+	"app/auth"
 	"app/db"
 	"app/ent"
 	"app/utils"
@@ -60,7 +61,7 @@ func (s *SigninServiceTestSuite) TestSignin() {
 
 	s.Run("success", func() {
 		res, err := s.service.Signin(s.ctx, targetUser.Email, "12345")
-		token, err := s.service.ValidateToken(res.Token)
+		token, err := auth.ValidateToken(res.Token)
 
 		s.Equal(targetUser.ID, res.UserID)
 		s.NotEmpty(token)
